fixed offsets

+++ Overlord/Overlord/Infrastructure/Messages/ResizeWindowMessage.cpp
@@ -2,7 +2,7 @@
 
 
 ResizeWindowMessage::ResizeWindowMessage(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam) : WinMessageBase(hwnd, msg, wParam, lParam) {
-	this->resizeTypeRequested = wParam;
+	this->resizeTypeRequested = (DWORD)wParam;
 	this->newHeight = LOWORD(lParam);
 	this->newWidth = LOWORD(lParam);
 }
